feature: wip separate leaderboards
