+++ apps/evm/src/pages/Fusion/components/Leaderboard/index.tsx
@@ -1,1 +1,2 @@
-export { Leaderboard } from './Leaderboard';
+export { AllUsersLeaderboard } from './AllUsersLeaderboard';
+export { QuestUsersLeaderboard } from './QuestUsersLeaderboard';
