+++ apps/evm/src/utils/api-client.ts
@@ -19,12 +19,13 @@ export type UserResponse = {
   data: any;
   created_at: Date;
   updated_at: Date;
-  leaderboardRank: { user_address: string; total_reward_points: number; rank: number };
+  leaderboardRank: { user_address: string; total_reward_points: number; total_quest_points: number; rank: number };
   depositStats: any[];
   totalUsdDeposited: number;
   withdrawStats: any[];
   harvested: { partner_name: string; partner_refcode: string; total_points: 'string' }[];
   quests_breakdown: Record<string, number>;
+  total_quest_points: string;
 };
 
 type LeaderboardResponse = {
@@ -57,9 +58,11 @@ type LeaderboardItem = {
   rank: string;
   deposit_owner: Address;
   total_points: string;
+  total_quest_points: string;
   username: string;
   referred_by?: string;
-  quests_breakdown: Record<string, number>;
+  quests_breakdown?: Record<string, number>;
+  points_breakdown?: Record<string, number>;
 };
 
 type TVLStats = {
